add xlnet

+++ ziyue_proj/pipeline.py
@@ -56,7 +56,7 @@ def model_fn(batch_size,model_name):
         features = {}
         with tf.variable_scope("input", reuse=tf.AUTO_REUSE):
             features["input_ids"] = tf.cast(100 * tf.placeholder(tf.float32, shape=(batch_size, 384)), tf.int32)
-            features["input_mask"] = tf.cast(100 * tf.placeholder(tf.float32, shape=(batch_size, 384)), tf.int32)
+            features["input_mask"] = tf.cast(100 * tf.placeholder(tf.float32, shape=(batch_size, 384)), tf.float32)
             features["segment_ids"] = tf.cast(100 * tf.placeholder(tf.float32, shape=(batch_size, 384)), tf.int32)
             features["start_positions"] = tf.cast(100 * tf.placeholder(tf.float32, shape=(batch_size,)), tf.int32)
             features["end_positions"] = tf.cast(100 * tf.placeholder(tf.float32, shape=(batch_size,)), tf.int32)
